Refactored SearchBar
